Crypto.Macaroon

Description

Pure haskell implementations of macaroons.

Warning: this implementation has not been audited by security experts. Do not use in production

References:

Macaroons: Cookies with Contextual Caveats for Decentralized Authorization in the Cloud http://research.google.com/pubs/pub41892.html

Time for better security in NoSQL http://hackingdistributed.com/2014/11/23/macaroons-in-hyperdex

Types

data Macaroon

Main structure of a macaroon

data Caveat

Caveat structure

type Secret = ByteString

Type alias for Macaroons secret keys

type Key = ByteString

Type alias for Macaroons and Caveat and identifiers

type Location = ByteString

Type alias for Macaroons and Caveat locations

type Sig = ByteString

Type alias for Macaroons signatures

Accessing functions

Macaroons

location :: Macaroon -> Location

Target location

identifier :: Macaroon -> Key

Macaroon Identifier

caveats :: Macaroon -> [Caveat]

List of caveats

signature :: Macaroon -> Sig

Macaroon HMAC signature

Caveats

cl :: Caveat -> Location

Caveat target location

cid :: Caveat -> Key

Caveat identifier

vid :: Caveat -> Key

Caveat verification key identifier

Create Macaroons

create :: Secret -> Key -> Location -> Macaroon

Create a Macaroon from its key, identifier and location

inspect :: Macaroon -> String

Inspect a macaroon's contents. For debugging purposes.

addFirstPartyCaveat :: Key -> Macaroon -> Macaroon

Add a first party Caveat to a Macaroon, with its identifier

Serialize

module Crypto.Macaroon.Serializer.Base64

Verify

module Crypto.Macaroon.Verifier