Crypto.Macaroon.Verifier

Description

Documentation

verify :: (Functor m, MonadIO m) => Secret -> [Caveat -> m (Maybe (Either ValidationError ()))] -> Macaroon -> m (Either ValidationError Macaroon)

Verify a Macaroon's signature and caveats, given the corresponding Secret and verifiers.

A verifier is a function of type MonadIO m => Caveat -> m (Maybe (Either ValidatorError Caveat)).

It should return:

Nothing if the caveat is not related to the verifier (for instance a time verifier is given an action caveat);

Just (Left (ParseError reason)) if the verifier is related to the caveat, but failed to parse it completely;

Just (Left (ValidatorError reason)) if the verifier is related to the caveat, parsed it and invalidated it;

Just (Right '()') if the verifier has successfully verified the given caveat

data ValidationError

Type representing different validation errors. Only ParseError and ValidatorError are exported, SigMismatch and NoVerifier are used internally and should not be used by the user

Constructors

ParseError String
A verifier had a parse error

ValidatorError String
A verifier failed